requirements.txt
binance-connector
python-dotenv
fastapi
uvicorn
requests
python-telegram-bot